fix loadout not loading on tutorial.
So, discovered a bit of an issue with the loadout system: it does not work properly if it's called on player spawn event (which is what firefight_tutorial did.

This is because SetLoadoutFile gets called twice at the same time, and rather than remove the items both of the loadouts overlap and cause the loadout to be a hybrid of the two loadouts.

If you choose to use player_loadoutchanger, call it with OnMapSpawn rather than a logic_eventlistener. There is a tad bit of delay with OnMapSpawn compared to the player_spawn event, which makes it work a lot better in this case.

Both maps have been updated to load the loadouts from the mapinfo file, rather than the entity, to prevent this happening in the future. FR looks for the ForcedLoadout value in a mapinfo file if the ForcedLoadout boolean is not on., which happens on map init. player_loadoutchanger should function if the mapinfo's ForcedLoadout is set, however there is a bug where the mapinfo will override the player_loadoutchanger. This will be fixed in the next beta update.

ugh